♻️ Move settings from core into base
