+++ lamindb/models/record.py
@@ -54,9 +54,9 @@ from lamindb_setup._connect_instance import (
     load_instance_settings,
     update_db_using_local,
 )
+from lamindb_setup.base._settings_store import instance_settings_file
 from lamindb_setup.core._docs import doc_args
 from lamindb_setup.core._hub_core import connect_instance_hub
-from lamindb_setup.core._settings_store import instance_settings_file
 from lamindb_setup.core.upath import extract_suffix_from_path
 
 from lamindb.base import deprecated
@@ -70,7 +70,7 @@ from lamindb.base.fields import (
 from lamindb.base.types import StrField
 from lamindb.errors import FieldValidationError
 
-from ..core._settings import settings
+from ..base._settings import settings
 from ..errors import (
     InvalidArgument,
     RecordNameChangeIntegrityError,
